refactor: split state machine events into input/output modules
